gcp_postgres_sandbox/quality_assurance_v2/requirements.txt
# CryptoPrism-DB Quality Assurance System v2 Dependencies

# Core Dependencies
sqlalchemy>=2.0.32
pg8000>=1.30.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.2.2
numpy>=1.26.4

# Notifications & Communication  
requests>=2.32.3
python-telegram-bot>=20.0

# AI Integration (Optional)
openai>=1.0.0
google-generativeai>=0.5.0

# Utilities
pathlib2>=2.3.0
typing-extensions>=4.0.0

# Development & Testing (Optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Logging & Monitoring
coloredlogs>=15.0.1